+++ TestFrameWorkWI2020C/src/main/java/com/example/testframeworkwi2020c/CoreSystem/CodeRunnerBackend.java
@@ -1,17 +1,9 @@
-package com.example.testframeworkwi2020c;
+package com.example.testframeworkwi2020c.CoreSystem;
 
-import com.example.testframeworkwi2020c.testSammlung.X_III_StringController_Controller;
-import com.example.testframeworkwi2020c.testSammlung.X_I_unfair_Dice_Controller;
-import com.example.testframeworkwi2020c.testSammlung.t04_OOP.T04_OOP_Ü05_Controller;
-import com.example.testframeworkwi2020c.testSammlung.t04_OOP.T04_OOP_Ü06_Controller;
-import com.example.testframeworkwi2020c.testSammlung.t04_OOP.T04_OOP_ü04_Controller;
 import javafx.util.Pair;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.net.URL;
@@ -23,92 +15,42 @@ import java.util.Optional;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 
+public class CodeRunnerBackend {
 
-public class TestController implements ITester{
-    @Override
-    public String startComparison(String aufgabenname, String jarFilePath) {
-        switch (aufgabenname) {
-            case "04_OOP_Ü04":
-                return testTelephone(jarFilePath);
-            case "04_OOP_Ü05":
-                return testFigure(jarFilePath);
-            case "04_OOP_Ü06":
-                return testTrain(jarFilePath);
-            case "X-I_unfair_dice":
-                return testWuerfel(jarFilePath);
-            case "X-III_StringController":
-                return testStringController(jarFilePath);
+    public static List<Pair<String, Object>> jarTest(String jarFilePath) throws Exception {
+        List<Pair<String, Object>> objects = new ArrayList<>();
+
+        // Laden der Jar-Datei
+        try (JarFile jarFile = new JarFile(jarFilePath)) {
+            // Erstellen eines URLClassLoaders mit der Jar-Datei
+            URL[] urls = {new URL("jar:file:" + jarFilePath + "!/")};
+            ClassLoader classLoader = URLClassLoader.newInstance(urls);
+            // Durchsuchen aller Klassen in der Jar-Datei
+            jarFile.stream()
+                    .filter(entry -> entry.getName().endsWith(".class"))
+                    .forEach(entry -> {
+                        try {
+                            String className = entry.getName().replace("/", ".").replace(".class", "");
+                            // Laden der Klasse
+                            Class<?> clazz = classLoader.loadClass(className);
+                            // Überprüfen, ob die Klasse abstrakt ist
+                            if (java.lang.reflect.Modifier.isAbstract(clazz.getModifiers())) {
+                                System.out.println("Die Klasse " + className + " ist abstrakt und kann nicht instanziiert werden.");
+                                return;
+                            }
+                            // Erstellen einer Instanz der Klasse
+                            Object instance = clazz.getDeclaredConstructor().newInstance();
+                            // Hinzufügen des Klassennamens (ohne Paketnamen) und der Instanz zur Liste
+                            String simpleClassName = clazz.getSimpleName();
+                            objects.add(new Pair<>(simpleClassName, instance));
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
+                    });
         }
-        return "";
+
+        return objects;
     }
-
-    public String testWuerfel(String jarFilePath) {
-        String textAnUser = "";
-        X_I_unfair_Dice_Controller testWuerfel = new X_I_unfair_Dice_Controller();
-        try {
-            textAnUser += testWuerfel.testFairDice(jarFilePath);
-            textAnUser += testWuerfel.testUnfairDice(jarFilePath);
-//            textAnUser += testWuerfel.testDealer(jarFilePath);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        return textAnUser;
-    }
-
-    public String testTelephone(String jarFilePath){
-        String textAnUser = "";
-        T04_OOP_ü04_Controller controller = new T04_OOP_ü04_Controller();
-        try {
-            textAnUser += controller.testSmartphone(jarFilePath);
-            textAnUser += controller.testMobile(jarFilePath);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return textAnUser;
-    }
-
-    public String testStringController (String jarFilePath) {
-        String textAnUser = "";
-        X_III_StringController_Controller controller = new X_III_StringController_Controller();
-        try {
-            textAnUser += controller.testStringController(jarFilePath);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return textAnUser;
-    }
-
-    // Tests für die Übungsaufgabe OOP_Übung05
-    // Weitergabe des "Text an User"
-    public String testFigure(String jarFilePath){
-        String textAnUser = "";
-        T04_OOP_Ü05_Controller controller = new T04_OOP_Ü05_Controller();
-        try {
-            textAnUser += controller.testCircle(jarFilePath);
-            textAnUser += controller.testQuadrat(jarFilePath);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return textAnUser;
-    }
-
-    // Tests für die Übungsaufgabe OOP_Übung06
-    // Weitergabe des "Text an User"
-    public String testTrain(String jarFilePath){
-        String textAnUser = "";
-        T04_OOP_Ü06_Controller controller = new T04_OOP_Ü06_Controller();
-        try {
-            textAnUser += controller.testZug(jarFilePath);
-            textAnUser += controller.testGueterzug(jarFilePath);
-            textAnUser += controller.testSpezialzug(jarFilePath);
-            textAnUser += controller.testPersonenzug(jarFilePath);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return textAnUser;
-    }
-
 
     public static Object invokeMethodByName(List<Pair<String, Object>> classObjects, String className, String methodName, Object[] methodArgs, Class<?>[] parameterTypes) {
         Optional<Object> optionalObject = classObjects.stream()
@@ -131,6 +73,7 @@ public class TestController implements ITester{
         }
         return null;
     }
+
     public static Object invokeMethodByName(List<Pair<String, Object>> classObjects, String className, String methodName, Object... methodArgs) {
         Optional<Object> optionalObject = classObjects.stream()
                 .filter(pair -> pair.getKey().equals(className))
@@ -161,7 +104,6 @@ public class TestController implements ITester{
         return null;
     }
 
-
     public static Object getObjectByClassName(List<Pair<String, Object>> objectList, String className) {
         for (Pair<String, Object> pair : objectList) {
             if (pair.getKey().equals(className)) {
@@ -179,16 +121,6 @@ public class TestController implements ITester{
             }
         }
         throw new IllegalArgumentException("Interface not found in the object's interfaces: " + interfaceSimpleName);
-    }
-
-    public static Class<?> getInterfaceFromList(File jarFile, String interfaceSimpleName) throws Exception {
-        List<Class<?>> interfaceList = extractInterfacesFromJar(jarFile);
-        for (Class<?> interfaceClass : interfaceList) {
-            if (interfaceClass.getSimpleName().equals(interfaceSimpleName)) {
-                return interfaceClass;
-            }
-        }
-        throw new IllegalArgumentException("Interface not found in the list: " + interfaceSimpleName);
     }
 
     public static List<Class<?>> extractInterfacesFromJar(File jarFile) throws Exception {
@@ -210,6 +142,17 @@ public class TestController implements ITester{
 
         return interfaces;
     }
+
+    public static Class<?> getInterfaceFromList(File jarFile, String interfaceSimpleName) throws Exception {
+        List<Class<?>> interfaceList = extractInterfacesFromJar(jarFile);
+        for (Class<?> interfaceClass : interfaceList) {
+            if (interfaceClass.getSimpleName().equals(interfaceSimpleName)) {
+                return interfaceClass;
+            }
+        }
+        throw new IllegalArgumentException("Interface not found in the list: " + interfaceSimpleName);
+    }
+
     private static List<String> getClassNamesFromJar(File jarFile) throws IOException, IOException {
         List<String> classNames = new ArrayList<>();
 
@@ -227,4 +170,6 @@ public class TestController implements ITester{
 
         return classNames;
     }
+
+
 }
