+++ TestFrameWorkWI2020C/src/main/java/com/example/testframeworkwi2020c/CoreSystem/HelloApplication.java
@@ -1,4 +1,4 @@
-package com.example.testframeworkwi2020c;
+package com.example.testframeworkwi2020c.CoreSystem;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -6,7 +6,6 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 public class HelloApplication extends Application {
