+++ TestFrameWorkWI2020C/src/main/java/com/example/testframeworkwi2020c/CoreSystem/ITester.java
@@ -1,4 +1,4 @@
-package com.example.testframeworkwi2020c;
+package com.example.testframeworkwi2020c.CoreSystem;
 
 public interface ITester {
     public String startComparison(String aufgabenname, String filepath);
