+++ TestFrameWorkWI2020C/src/main/java/com/example/testframeworkwi2020c/testSammlung/t04_OOP/ü04/SmartphoneTester.java
@@ -1,7 +1,7 @@
 package com.example.testframeworkwi2020c.testSammlung.t04_OOP.ü04;
 
-import com.example.testframeworkwi2020c.CodeRunnerBackend;
-import com.example.testframeworkwi2020c.TestController;
+import com.example.testframeworkwi2020c.CoreSystem.CodeRunnerBackend;
+import com.example.testframeworkwi2020c.CoreSystem.TestController;
 import javafx.util.Pair;
 
 import java.io.ByteArrayOutputStream;
@@ -24,7 +24,7 @@ public class SmartphoneTester {
     public Boolean testClick() throws Exception {
         objectList = CodeRunnerBackend.jarTest(jarFilePath);
         System.setOut(new PrintStream(outputStreamCaptor));
-        TestController.invokeMethodByName(objectList,className,"dial");
+        CodeRunnerBackend.invokeMethodByName(objectList,className,"dial");
         System.setOut(standardOut);
         if (outputStreamCaptor.toString().contains("Klick")) {
             return true;
