+++ TestFrameWorkWI2020C/src/main/java/com/example/testframeworkwi2020c/testSammlung/t04_OOP/ü05/CircleTester.java
@@ -1,8 +1,7 @@
 package com.example.testframeworkwi2020c.testSammlung.t04_OOP.ü05;
 
-import com.example.testframeworkwi2020c.CodeRunnerBackend;
-import com.example.testframeworkwi2020c.TestController;
-import com.example.testframeworkwi2020c.testSammlung.t04_OOP.ü04.SmartphoneTester;
+import com.example.testframeworkwi2020c.CoreSystem.CodeRunnerBackend;
+import com.example.testframeworkwi2020c.CoreSystem.TestController;
 import javafx.util.Pair;
 
 import java.io.ByteArrayOutputStream;
@@ -25,7 +24,7 @@ public class CircleTester {
         objectList = CodeRunnerBackend.jarTest(jarFilePath);
 
         System.setOut(new PrintStream(outputStreamCaptor));
-        TestController.invokeMethodByName(objectList,className,"getArea");
+        CodeRunnerBackend.invokeMethodByName(objectList,className,"getArea");
         System.setOut(standardOut);
         if (outputStreamCaptor.toString().contains("pi * r hoch 2")) {
             return true;
@@ -38,7 +37,7 @@ public class CircleTester {
         objectList = CodeRunnerBackend.jarTest(jarFilePath);
 
         System.setOut(new PrintStream(outputStreamCaptor));
-        TestController.invokeMethodByName(objectList,className,"getCircumference");
+        CodeRunnerBackend.invokeMethodByName(objectList,className,"getCircumference");
         System.setOut(standardOut);
         if (outputStreamCaptor.toString().contains("2*pi*r")) {
             return true;
