+++ TestFrameWorkWI2020C/src/main/java/module-info.java
@@ -18,8 +18,8 @@ module com.example.testframeworkwi2020c {
     requires org.junit.platform.commons;
     requires org.junit.platform.launcher;
 
-    opens com.example.testframeworkwi2020c to javafx.fxml;
-    exports com.example.testframeworkwi2020c;
     exports com.example.testframeworkwi2020c.testSammlung;
     opens com.example.testframeworkwi2020c.testSammlung to javafx.fxml;
+    exports com.example.testframeworkwi2020c.CoreSystem;
+    opens com.example.testframeworkwi2020c.CoreSystem to javafx.fxml;
 }
